Changed Calendar Event Input from dialog to Fragment style
