+++ app/src/main/java/com/example/bellyfull/modules/PersonalisedNotifications/Fragments/BlankFragment.java
@@ -1,0 +1,21 @@
+package com.example.bellyfull.modules.PersonalisedNotifications.Fragments;
+
+import android.os.Bundle;
+
+import androidx.fragment.app.Fragment;
+
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import com.example.bellyfull.R;
+
+
+public class BlankFragment extends Fragment {
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_blank, container, false);
+    }
+}
