+++ app/src/main/res/drawable/rounded_corner.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <corners
+        android:topLeftRadius="60dp"
+        android:topRightRadius="60dp" />
+    <solid android:color="@android:color/white" />
+</shape>
