+++ app/src/main/res/layout/dialog_event_input.xml
@@ -3,9 +3,11 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
+    xmlns:tools="http://schemas.android.com/tools"
     android:background="@color/white"
     android:orientation="vertical"
-    android:padding="14dp">
+    android:padding="14dp"
+>
 
 
     <EditText
